requirements.txt
numpy==1.17.2
tensorflow-cpu==2.3.0
streamlit==0.72.0
Pillow==8.0.1
